declare function for getWindowTitle (see below)
The way it ran for several years was broken. Till now I do not understand why. The problem was that the variable `$title` which was inserted into the json map was empty. I have never had it before that a defined variable which definitely has content is returned as empty. I guess the very big conditional was too much for the database. However I created a function which does exactly the same and passed that to the json map. Now, the behavior is as expected, even if I still don't understand the problem.