create setup.md
this commit adds a detailed setup description for Edirom-Online and a data package. It is more detailed than in the readme.md to make every step as clear as possible, since the readme.md instructions are fine for advanced developers that need only a short and quick description. 
thanks to @feuerbart for capturing this during the edirom-summer-school 2024. i corrected only some minor typos and changed only a little bit regarding the structure.

Refs https://github.com/Edirom/Edirom-Online/issues/415