update <front/>, delete Preparation
this commit updated the <front> text with some simplifications of sentences. Preparation chapter is deleted, because of two reasons. first, it is totally outdated and describes how to install the software from a DVD. Second, how to setup an Edirom locally is described in the docs-directory of the software-repository, the USER help is not an appropriate place for this.

refs #471